requirements.txt
discord.py==2.3.2
Flask==3.0.3
aiohttp==3.9.5
requests==2.32.3
Pillow
python-dotenv
PyNaCl
pydub
